refactor : #245 <Coupon> AuthenticatedUser 적용 및 인터페이스 생성
- 인터페이스를 생성하고 구현체 적용
- AuthenticatedUser로 role 체크
